Move progress_tracker() to util.py file in order to be used globally
